Fix bugs in and improve performance of LocalTypeFieldExtractor
This PR fixes several bugs in the LocalTypeFieldExtractor GetAllNodes
method. It also drastically improves the performance of the code. Tests
and a benchmark for the updated code were added in #297. Details of the
node extraction algorithm are described inline.

See the updated and uncommented tests for details about the fixed bugs
(and well as the fixed issues linked below).

The benchmark for the updated code is as follows:
```
$ cd pkg/engine/plan
$ go test -bench=. -run=xxxx
goos: darwin
goarch: amd64
pkg: github.com/jensneuse/graphql-go-tools/pkg/engine/plan
BenchmarkGetAllNodes-8  53476  22360 ns/op  10804 B/op  173 allocs/op
PASS
ok  github.com/jensneuse/graphql-go-tools/pkg/engine/plan  1.872s
```

This is a 500x speedup over the previous implementation (the prior
benchmark results can be found in #297).

Fixes #293 and #294.