Treat a union intersecting an interface symmetrically from interface | union.
When expanding a named fragment, one of the cases to consider is that
the named fragment has union type, but the enclosing type is an
interface type.  (e.g. `user { ...data } fragment data on SomeUnion {...}`,
where `user` has an interface type.  fragment_spread_inlining.go already
treats this case correctly.

However, there's a symmetrical case where the enclosing type is a
union, and the naamed fragment has an interface type.  According to
the spec, this should be treated identically, but
fragment_spread_inlining.go was not considering this case.  With this
PR it is.

Fixing this turned up another problem in ast_node.go:
NodeFragmentIsAllowedOnInterfaceTypeDefinition properly allowed the
fragment on union types when appropriate, but in the symmetrical case,
NodeFragmentIsAllowedOnUnionTypeDefinition _never_ allowed the
fragment on interface types.  Now it uses the same condition as the
other case, allowing it when there's a non-zero intersection of the
two types.