Avoid printing the same variable twice for a query.
Before this change, if a variable was used in a query both at the top
level and inside a recursive structure, it would be included twice in
the query:
```
query($name: String!, $name: String!){user(name: $name){normalized(data: {name: $name})
```

After this change, it only prints it once.  I think the previous code
location was just a mistake.