Implement federation of abstract fields
Fixes #328.

(Note: this change looks huge, but approximately 1800 of the lines are
tests and 150 of the lines are comments).

This PR implements federation of abstract fields. Specifically, it adds
support for the situation where a field with an interface or union type
has concrete types that are owned by another service (or at least have
selected fields owned by another service).

Take, for example, the query:
```
query {
  user {
    username
    pets {  # interface Pet
      name
      ... on Dog {
        dogField
      }
      ... on Cat {
        catField
      }
    }
  }
}
```

In this example "user" and "pets" are both resolved by the "user"
service and "name", "dogField" and "catField" are resolved by the "pet"
service. The issue is that the "pets" field is owned by one service but
the fields of the Pet type are owned by another service.

The major changes are:
- During planning, the first concrete type of an interface is used to make planning decisions. (This requires that a field with a given name is owned by the same service for all concrete types.)
- Also during planning, the first concrete type of an interface is used for field configuration. The same requirement as above applies.
- The federation root is added to the upstream opertion the first time an inline fragment, selection set or field is visited below the planner parent path.
- A "__typename" selection is always added to interface field selection sets, it case it's needed for federation (this isn't necessary for unions because union selection sets have explicit inline fragments).
- The "__typename" variable is now dynamic in representations when the root type is abstract.
- Multiple inline fragments are added to the federation root, one for each concrete type (this only applies to interface, since unions come with their own inline fragments).
- All sub-selections are added to the concrete inline fragments, and then fragments with invalid type conditions or empty fragments are pruned during upstream operation normalization.