Implement copying of query document AST nodes
This PR implements copying of AST nodes used in query documents. Copying
will be used when normalizing operations for federation. Selection sets
will be copied when "expanding" inline fragments and inserting inline
fragments for fields with interface types. See the plan in #328 for
details.

I'm not sure everything that's being copied needs to be copied. For
example, it seems like it would be odd for someone to modify a value,
but I copy everything except byte slice references to be one the safe
side.

Copying could also be utilized when inlining fragments, if we wanted.
Right now there's a big warning on ReplaceFragmentSpread about field
aliasing (and it isn't clear to be that field de-duplication fully fixes
the aliasing).