Fix `addVariableDefinitionsRecursively` to handle arrays inside objects.
`resolveNestedArgumentType` wasn't returning the proper type when the
argument was an array inside an object.  There were two reasons for
this: first, the proper object field wasn't being carried through the
recursive addVariableDefinitionsRecursively() call that was made in
the "list" case, and second, nothing was looking through the
"array-of" to get the type of the underlying object.

The fix is just to get rid of that use of `resolveNestedArgumentType`
directly.  There's no need to try to figure out the type of the
variable from parsing the complex nesting; we already know the type of
the variable from the original document!  So now we just use that.