[TT-6110] Allow list for field based permissions implement allow block full types with asterisk (#295)
PR for
[TT-6110](https://tyktech.atlassian.net/jira/software/c/projects/TT/boards/26?modal=detail&selectedIssue=TT-6110)

Initially, I thought it was good to use an empty fields list to implement that feature but the current implementation validates the input if the fields list is empty. So in order to avoid breaking the API and its behavior, I decided to use an asterisk(*) character to denote all fields of a type are allowed or blocked.

See the following snippets:

```go
	FieldRestrictionList{
		Kind: AllowList,
		Types: []Type{
			{
				Name:   "Query",
				Fields: []string{"hero"},
			},
			{
				Name:   "Character",
				Fields: []string{"*"},
			},
		},
	}
```

The above sample allows all fields of the `Character` type.

```go
	FieldRestrictionList{
		Kind: BlockList,
		Types: []Type{
			{
				Name:   "Character",
				Fields: []string{"*"},
			},
		},
	}
```

The above sample restricts all fields of the `Character` type.