[TT-6110] Fix inconsistent behavior of the asterisk operator (#299)
This PR fixes the bug described here:
https://tyktech.atlassian.net/browse/TT-6110?focusedCommentId=28231

The bug is caused by the mislocation of the asterisk check and iterating
over a map. Now it checks the asterisk character first for a requested
type, then tries to find a disallowed field if there is any.

Also, I would like to clarify the usage of this allow-list feature. The
following config snippet allows to access all current and future fields
of the `Query` type. But the other types and their fields are still
disallowed to access. So the asterisk operator doesn't work recursively.

```json
"allowed_types": [
                {
                    "name": "Query",
                    "fields": ["*"]
                }
]
```

The following configuration snippet allows to access all current and
future fields of the `Query` type. It also allows to access the `name`
field on both `Country` and `Continent` types.

```json
"allowed_types": [{
			"name": "Query",
			"fields": ["*"]
		},
		{
			"name": "Country",
			"fields": ["name"]
		},
		{
			"name": "Continent",
			"fields": ["name"]
		}
	]
```

In short, the asterisk(*) operator only allows access to fields of a
specific type. It doesn't work recursively.