[TT-6669] Strict version of IsIntrospectionQuery (#300)
This PR proposes a solution for the problem described here:
[TT-6669](https://tyktech.atlassian.net/jira/software/c/projects/TT/boards/26?modal=detail&selectedIssue=TT-6669)

After an initial investigation, I found that the original author
deliberately implemented the feature I described in the ticket.

In brief, the following query is an introspection query:

```graphql
{
  __schema {
    types {
      name
    }
  }
}
```

`IsIntrospectionQuery` returns true for this and this is the expected
behavior. But if we add an additional field to the above query,
`IsIntrospectionQuery` returns false and this behavior breaks the
granular access policy of Tyk Gateway. See the related ticket on Jira:
[TT-6568](https://tyktech.atlassian.net/browse/TT-6568)

```graphql
{
  __schema {
    types {
      name
    }
  }
  countries { name }
}
```

So the above query returns the introspection response and the name of
all countries worldwide even if a user deliberately disabled GQL
introspection on this API.

It looks like this is the expected behavior for `IsIntrospectionQuery`
because I found the following tests in
`TestRequest_IsIntrospectionQuery`:

```go
t.Run("schema introspection query with additional non-introspection fields", run(nonSchemaIntrospectionQueryWithAdditionalFields, false))
t.Run("type introspection query with additional non-introspection fields", run(nonTypeIntrospectionQueryWithAdditionalFields, false))
```

In order to protect the current behavior and prevent breaking the other
users' code, I decided to add a stricter version of
`IsIntrospectionQuery`. It is called `IsIntrospectionQueryStrict`.
Basically, it tries to find the `__schema` and `__type` fields in a
given query by scanning the operation and fragment/inline fragment
definitions.

I used the same test cases to verify its function. I also never touched
the business logic of `IsIntrospectionQuery`. I just split it into
smaller chunks to reuse some code.