[TT-6798] Reconnection feature is broken in subscription (#309)
This PR fixes the wrong handling of connection errors in the GQL
subscription. Previously, it was hiding the connection errors due to
this faulty behavior, it was failing to trigger the reconnection
mechanism. It was waiting forever in an uninterruptible and unobservable
state.

I added integration tests to simulate an upstream failure. Basically,
the tests kill the underlying TCP socket and trigger the reconnection
mechanism, and return the error messages.

For example, if I kill the upstream service at some point it returns the
following error messages:

```json
{"id":"1","type":"data","payload":{"errors":[{"message":"unable to resolve","locations":[{"line":2,"column":3}],"path":["getUserNotifications"]},{"message":"failed to get reader: failed to read frame header: EOF"}],"data":null}}
{"id":"1","type":"error","payload":[{"message":"failed to WebSocket dial: failed to send handshake request: Get \"http://localhost:4204/query\": dial tcp [::1]:4204: connect: connection refused"}]}
{"id":"1","type":"error","payload":[{"message":"failed to WebSocket dial: failed to send handshake request: Get \"http://localhost:4204/query\": dial tcp [::1]:4204: connect: connection refused"}]}
{"id":"1","type":"error","payload":[{"message":"failed to WebSocket dial: failed to send handshake request: Get \"http://localhost:4204/query\": dial tcp [::1]:4204: connect: connection refused"}]}
```

This client may choose a STOP message to halt the subscription instead
of waiting to reconnect to the upstream.

By the way, the GQL resolver tries to resolve the first error message. I
think the `type` should be `error` instead of `data`. Is this a bug or a
feature? cc: @pvormste

```json
{"id":"1","type":"data","payload":{"errors":[{"message":"unable to resolve","locations":[{"line":2,"column":3}],"path":["getUserNotifications"]},{"message":"failed to get reader: failed to read frame header: EOF"}],"data":null}}
```