[TT-6419] Querying for __typename in any graph causes a 502 error (#323)
Currently, when you send the following query for both proxy-only and UDG
APIs, the GW panics and the user gets 502 for the query.

```graphql
query {
    __typename
}
```

The actual error message is the following:

```json
{
    "errors": [
        {
            "message": "unable to resolve",
            "locations": [
                {
                    "line": 2,
                    "column": 5
                }
            ]
        }
    ],
    "data": null
}
```

The `data` is null because no data source is attached to the
`__typename`. It's evaluated by this block:
https://github.com/TykTechnologies/graphql-go-tools/blob/c2dd74b524578d0653cf7652efba56a19304ebae/pkg/engine/plan/plan.go#L519

Initially, I thought that it would be enough to proxy the request to the
upstream. It works for the proxy-only mode but it doesn't work for UDG.

Consider the following example. I have a dummy REST service that returns
`{"temp": random-number}` for every request. It doesn't speak GraphQL.

```graphql
query {
    __typename
    weather {
        temp
    }
}
```

UDG returns the following response for this query because the
`__typename` doesn't exist in the response.

```json
{
    "errors": [
        {
            "message": "unable to resolve",
            "locations": [
                {
                    "line": 2,
                    "column": 5
                }
            ]
        }
    ],
    "data": null
}
```

I realized that the `__typename` meta-field should be maintained by the
library directly when required. The most simple and feasible way to
export the type name from planner to resolver is to add a new field to
the `resolver.Field` struct. The other option is to add a type name
field to `resolver.Object` but I couldn't find a way to get the root
type name. In addition to that, it breaks many tests in resolve and
planner packages.

The response after the fix was applied:

```graphql
query {
    __typename
    weather {
        __typename
        temp
    }
}
```

```json
{
    "data": {
        "__typename": "Query",
        "weather": {
            "__typename": "Weather",
            "temp": 44
        }
    }
}
```