[TT-7468] Configuring UDG Schema based on info parsed from AsyncAPI spec (#329)
PR for
[TT-7468](https://tyktech.atlassian.net/jira/software/c/projects/TT/boards/26?modal=detail&selectedIssue=TT-7468)

`ImportAsyncAPIDocumentByte` has been added to produce a GraphQL schema
from a parsed AsyncAPI document. The implementation uses
`introspection.JsonConverter` to build a GraphQL schema. Introspection
was used as an intermediate format between GQL and AsyncAPI documents.
This way, we avoid implementing a clone of `introspection.JsonConverter`
for AsyncAPI.

A few things to consider:

* `introspection.__TypeKind` is an unexported type but the
implementation requires it. As a workaround, I added a function that
returns `introspection.TypeRef` directly. We can export `__TypeKind`
from the introspection package but this is a breaking change.
* The parsed AsyncAPI document uses built-in maps to store data. We
iterate over these maps to create an introspection JSON. Because the
iteration order in a map is random, every iteration produces a different
GQL schema. In order to overcome this problem, I sort the slices by name
(it can be any field).

`operationId` is required to construct the `Subscription` type.
Otherwise, the converter produces the following result:

```graphql
type Subscription {
   : PaymentFailed
   : PaymentSucceeded
   : UserRegistered
   : UserSubscribed
   : UserUnsubscribed
}
```

The following sample contains the `operationId` for subscriptions. 

```yaml
  payment-failed:
    subscribe:
      operationId: paymentFailedSub
      description: The payload of failed payment
      message:
        $ref: "#/components/messages/payment"
```

The result looks like the following:

```graphql
type Subscription {
    paymentFailedSub: PaymentFailed
    paymentSucceededSub: PaymentSucceeded
    userRegisteredSub: UserRegistered
    userSubscribedSub: UserSubscribed
    userUnsubscribedSub: UserUnsubscribed
}
```

[TT-7468]:
https://tyktech.atlassian.net/browse/TT-7468?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ