[TT-8528] Improper handling of non-nullable fields in OpenAPI (#350)
PR for [TT-8528](https://tyktech.atlassian.net/browse/TT-8528)

Now, we can handle non-nullable fields properly. There was also a minor
bug in a function that generates mutation type. It was swallowing the
parameters if a request body defined for the same operation id. This bug
has also been fixed.

[TT-8528]:
https://tyktech.atlassian.net/browse/TT-8528?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ