fix full type details not being able to be updated (#395)
This PR fixes an issue where it could happen that the OAS converter
picks up a full type without all the details before reaching the full
definition. This caused the converter to not pick up descriptions.

To fix this I introduced an details cache that only tracks if a detail
is already stored or not.