[TT-11071] OAS converter fails when return type is JSON (#416)
Tyk GW uses `github.com/iancoleman/strcase v0.3.0` but graphql-go-tools
uses version `v0.2.0`. The OpenAPI translator uses the `strcase.ToCamel`
method to process type names. This method in the newer version converts
`JSON` to `Json`, while the older version doesn't do that. It simply
returns the same string.

This PR fixes this problem and checks the type name before further
processing. We don't need to call `ToCamel` method if the type name is a
predefined scalar type.

We don't catch this error with our test suite because when
graphql-go-tools is imported from Tyk GW, the newer version of the
dependency is used. It may be a good idea to upgrade
`github.com/iancoleman/strcase` to `v0.3.0` from `v0.2.0` for
compatibility. I do not prefer to do that in this PR because different
components of the library depend on it. It should be handled in a
different story.