[TT-11249] OAS-to-GQL translator adds the same field many times (#423)
This PR fixes bad handling of responses if an HTTP status code range and
explicit code definition are used at the same time. The OpenAPI
specification is clear about this:

> To define a range of response codes, you may use the following range
definitions: 1XX, 2XX, 3XX, 4XX, and 5XX. If a response range is defined
using an explicit code, the explicit code definition takes precedence
over the range definition for that code.

Now, the OpenAPI-to-GraphQL translator picks the explicit code
definition over the range definition.

Suppose the OpenAPI document contains multiple possible successful
response objects (HTTP code 200-299 or 2XX). Only one can be chosen. It
selects the first response object with the successful status code
(200-299). The response object with the HTTP code 200 will be selected
if defined. This behavior is borrowed from IBM/openapi-to-graphql tool.

This PR also removes redundant and useless code in response handling. 

The changes will be copied to the `v2` folder after approval.