[TT-11295] Gateway Panics when arguments are not defined in operation section (#425)
This PR fixes TT-11295. The problem was not related to the Persist GQL
feature. The library fails to validate a specifically crafted query in
which variables are not defined in the operation.

An example query: 

```graphql
query Analytics ($startMonth: String, $startYear: String!, $endDay: String!, $endMonth: String, $endYear: String!){
  analytics(
    EndDay: $endDay
    EndMonth: $endMonth
    EndYear: $endYear
    StartDay: $startDay
    StartMonth: $startMonth
    StartYear: $startYear
  ) {
    data {
      error
      hits
      success
      id {
        alias
        key
        additional_data {
          data {
            meta_data {
              department
            }
            tags
          }
        }
      }
    }
  }
}
```

`$startDay` is not defined in the operation `Analytics`. The library
panics when you try to run this query. With this PR, it returns the
following error:

```json:
{
  "message": "Variable \"$startDay\" is not defined by operation \"Analytics\"",
  "locations": [
    {
      "line": 6,
      "column": 15
    }
  ],
  "path": [
    "query",
    "analytics"
  ]
}
```