should resolve with a nested lonely data source (TT-9884) (#429)
This PR fixes an issue when a nested data source is lonely.

It is built on top of @buraksezer 's research.

How does it work?

The GraphQL data source planner has been slightly modified and will
before printing the upstream query check for fields that require
selections.

Example:
```
{ countries }
```

During this check, it will realize that `countries` is missing the
selections and will therefore add the `__typename` selection field. This
will result in following upstream query:
```
{ countries { __typename } }
```

With that the upstream is valid again.