fix(edfs): filter only when type matches (#805)
This PR ensures that we respect the value type when comparing the
argument value with the event payload. Before this PR, we treated "1" as
1 and vice versa, as well as "true" = true.

Changes in a nutshell: When we have one segment we can extract the type.
If we have more than one it will always result in a byte value
comparison. In case of array values, we can always get the type and
compare it with the argument type.