[TT-9864] Optimize the creation usage of AST documents (#431)
PR for [TT-9864](https://tyktech.atlassian.net/browse/TT-9864).

After reading a lot of code, trying different solutions and measuring
them, I decided to go with the simplest solution.

I used `sync.Pool` for reusing `ast.Document` instances and added a
`Cleanup` method to `Request` struct to recycle `ast.Document` structs
**if** it's possible.

The library caches the plans for queries but before caching or deciding
to use a plan from the cache, it creates a new `ast.Document`, parses
the query, normalizes and validates it against the schema. If a plan is
cached, resetting and putting `ast.Document` to the pool can lead to
subtle bugs because the cached plan still uses it for the incoming
queries.

I added a boolean field to `Request` struct to mark points out the
request will be processed by a plan from the cache. If so, we don't need
the `ast.Document` in the `Request` struct anymore and we can recycle
it. `Cleanup` method does this job. I plan to use `Cleanup` method from
Tyk GW. I tried to run the CI pipeline with a dummy PR and it worked.
See this: https://github.com/TykTechnologies/tyk/pull/6345/files#

Because `Cleanup` method is only called from Tyk GW, it should not has
any side effects on the tests in the library.

`ast.NewDocument` method has been called from different places in the
library but recycling `ast.Document` in GraphQL data source and
federation code leads to bugs and breaks the tests.


[TT-9864]:
https://tyktech.atlassian.net/browse/TT-9864?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ