fix: processing websockets dials asynchronously (#831)
This PR improves the way we handle websocket dialing. Establishing a
websocket connection with the origin can take up to 30s before it times
out. For this reason, dialing should be done asynchronously to not
interfere with other subscriptions. Additionally, blocking all sockets
has led to starvation of the epoll event loop, which results in missing
connections to be reported for deletion.

This PR also deletes the deduplication logic of websocket connections in
the client because this is done already in the router.

---------

Co-authored-by: Sergiy 🇺🇦 <818351+devsergiy@users.noreply.github.com>