feat: improve resolve performance by solving merge abstract nodes in postprocessing (#826)
Previously, we've generated too many fields for abstract nodes. This
meant that after resolving, we had to deduplicate and merge some fields.
This new approach solves the problem in a post processing step as part
of the planning. This means that the runtime does less work during
resolving, and a little bit more work during planning, which gets cached
anyways.

---------

Co-authored-by: Sergiy 🇺🇦 <818351+devsergiy@users.noreply.github.com>