fix: argument and variable validation during execution (#902)
When using `engine.Execute` from the `execution` package, I generally
expect that all inputs will be validated before the operation is
actually executed against the data source. As discussed in #885, this
turned out to not be the case for:

- Input arguments that are specified but are invalid when compared to
their schema
- Input arguments that are an input object type, where one of the fields
of the input object is invalid when compared to its schema
- User-supplied variables that are invalid when compared to the variable
definition in the operation

After much research, I believe there are two core issues:

- The default normalization options include `WithExtractVariables`,
which converts arguments to variables. Since normalization happens
before validation, a document with invalid arguments is normalized to a
_valid_ document with invalid variables. Thus validation of the document
passes erroneously.
- Variable validation is not performed at all. Even if it were, it would
give incorrect messages about invalid arguments - because they would
have been converted to variables using assigned names rather than the
original argument name and path.

See #885 for examples of both cases.

The solution is to break up the normalization process into multiple
stages.
1. First, normalize with all the default options _except_
`WithExtractVariables`.
 2. Next, validate the operation document
3. Then validate the variables the user supplied with the original
request
4. Finally, normalize again, extracting document arguments to variables
so the execution plan caching continues to work as originally designed.

After applying these changes, two of the existing tests failed. On
closer inspection, they were actually false-positive tests for the very
problems I was experiencing. One of them was allowing null values
through to the data source where it should have errored, and the other
was being fed a single integer to a variable that was expecting a list
of integers. I updated those tests to the correct behavior.

Let me know if you need any adjustments or more tests added.  Thanks. 🙂

---------

Co-authored-by: Sergiy 🇺🇦 <818351+devsergiy@users.noreply.github.com>