fix(subscriptions): avoid deadlock on race for tools (#933)
We already limit the number of concurrent subscription updates through a
semaphore. Removing the race around tools removes a deadlock when many
subscriptions 7K+ are trying to close and update. This also impacted
regular http requests since the getTools was shared. I tested the update
with 10K subscriptions and couldn't find any impact on memory usage.