feat: make multipart heartbeat configurable (#1006)
Previous work for multipart added a timer for regular heartbeat messages
which could be set to keep the subscription connection alive. We had
configured it to always be 5 seconds, but that caused friction in our
cosmo tests which wanted to verify that the heartbeat would be sent.
This PR allows for it to be configurable via options, so that we can
change it in our tests and save a few seconds.