feat(engine): mutex free subscription handling (#1076)
This PR removes all mutexes needed and uses channels to synchronize
access in subscription handling. Every subscription spawns a single
goroutine where the fetch and resolve process is run (connection write).
Passing a task to a subscription is done on a buffered channel and on
the event loop (single threaded). In that way, we can guarantee that we
send messages in the order they were received, no matter how high the
concurrency is. Before, it could happen that events arrived out of
order. By using a buffered channel, we can increase the throughput of
how many updates can be offloaded to a subscription worker while other
subscriptions can continue. This implementation make use of the natural
back-pressure behaviour of go channels. A `MaxSubscriptionFetchTimeout`
is added to ensure that a single subscription fetch can't block the
event loop forever. This timeout is configurable and set to `30s` by
default.

Semaphores have been deleted as well because we now naturally limit the
throughput by writing to subscriptions sequentially via channels.
Additionally, heartbeat handling is done in the worker goroutine of
every subscription. This simplified the implementation and removed
another synchronization need.

We also massively improved the way synchronous subscriptions are
managed. There is no difference in handling anymore, except that for
SSE/Multipart, we wait until the client request is done or the
subscription is completed and drained.

I did a few benchmarks with 200 concurrent subscriptions and around 10
triggers. Writing to all subscriptions was still pretty fast (<1ms, ~3ms
Max). The next iteration of the implementation should focus on
leveraging concurrency better. We should handle triggers individually so
that the blast radius of a fast producer and slow writer and can't
impact others. Today, a fast producer will result in increased latency
but more stable CPU utilization due to backpressure.

Fixes ENG-6513

---------

Co-authored-by: Jens Neuse <jens.neuse@gmx.de>