fix(subscription): never try to send on blocked channel when subscription was completed (#1100)
This PR modifies the approach to synchronizing the shutdown of a
subscription. Instead of using a separate channel, it now closes the
`workChan` and terminates the goroutine only after all events have been
produced. The final close event will terminate the subscription worker
goroutine. The biggest improvement of this change is that closing the
channel is done on the same goroutine where subscription are managed.
This should eliminate race or deadlock conditions.