feat(subscriptions): user proper frame timeout (#1135)
This PR ensures that we use a proper frame timeout. Before we used the
`readTimeout` as frame timeout which is usually 50x higher than a
recommended value for the frames. This has result in conflicts with
other timeouts.

I tested this with a demo setup for 2h where a subgraph writes events
every 2minutes.