feat: remove intermediate buffer from ResolveGraphQLResponse (#1137)
Improves performance once https://github.com/wundergraph/cosmo/pull/1795
is merged, as that PR moves the intermediate buffer from engine into the
router.

This is technically a behaviourally breaking change, as `writer` could
be written to in `err` case, before it would not have been. Our tests
all still pass, which means we weren't asserting the previous behaviour,
but some consumer could theoretically depend on it as the router did.