feat: use new negation regex matching for matching connections (#1161)
In the router we use negation in regex expression matching via a
negation flag, since we also use this logic when creating connections,
and we need to assign connections correctly, we apply this logic in the
engine.

**Tests**
Add unit tests for checking the hash. There used to be tests that tests
if a connection is reused, but it has been removed with certain changes.
I would need more time to go through the engine to figure out how to
restructure the tests (and why they were removed in the first place)
https://github.com/wundergraph/graphql-go-tools/commit/3ec0ebef5a8ab6133b411ab009fe8da289701c5f#diff-2a8da55a93918e23bbe398ea85eb27a43292dd9cc3090eea26264094dc28931bL355-L365