fix: support different kinds of close, correct client unsubscribe behaviour (#1174)
In #1158, I added a new path for handling server closure, but as an
unintended side effect some scenarios that were previously silently
failing such as complete-after-error in subscription init events became
obviously broken.

This PR follows up on this by adding options for the _kind_ of closure
the caller wants, including 3 to start:

- "normal" = the connection is being closed normally and intentionally,
such as in the case of connection failure
- "downstream service error" = the subgraph closes the connection before
completing
- "going away" = used during router shutdown to indicate the router
itself is shutting down

It also:
- renames many instances of "Shutdown" to "Close" and "Done" to
"Complete" for consistent terminology through the router and engine for
subscriptions.
- corrects behaviour where we were not distinguishing between removing
clients for different reasons and always sending complete as if it was
because the client sent complete, now handles complete separately