fix: check that object claims to implement interface
Fragment valdidation layer calls the
"NodeFragmentIsAllowedOnInterfaceTypeDefinition" function to validate
that fragment node is allowed in the interface. But it was not checking
that the node actually claimed to implement the interface, only that
the node contained all the fields from interface.

I have added this additional check by calling to the
"NodeImplementsInterface" function.

Also tests were made more strict by adding specific errors to the
invalid cases.