fix: use NodeFragmentIsAllowedOnNode to check fragmentSpread (#1223)
This fixes following false positive error in the "interfaces" benchmark:

fragment spread: fragment catFrag must be spread on type Cat and not
type CatOrDog

Previously rule fragment check was trying to verify that the catFrag
fragment (of a Cat type) matches completely the name of EnclosingType
(CatOrDog). Instead I propose to use the "NodeFragmentIsAllowedOnNode"
method to check if fragment can be applied to the enclosingType that can
be union (our case).

Benchmark query:

    query conflictingDifferingResponses {
	    catOrDog {
		    ...catDogFrag
	    }
    }
    fragment catDogFrag on CatOrDog {
	    ...catFrag
	    ...dogFrag
    }
    fragment catFrag on Cat {
	    someValue: meowVolume
    }
    fragment dogFrag on Dog {
	    someValue: barkVolume
    }