fix: check that object claims to implement interface (#1235)
Fragment valdidation layer called the
"NodeFragmentIsAllowedOnInterfaceTypeDefinition" function to validate
that fragment node was allowed in the interface. But it was not checking
that the node actually claimed to implement the interface, only that the
node contained all the fields from interface.

I have replaced the checking of fields with the check if 
an object claims to implement an interface using the 
"NodeImplementsInterfaceNode" method.

Also tests were made more strict by adding specific errors to the
invalid cases.