feat: add support of field selection reasons extensions (#1282)
The PropagateFetchReasons option of Resolver enables router to send to
subgraphs an extension that includes reasons why fields where included
in the request.

This data is sent only for the fields marked with
`@openfed__requireFetchReason` in the schema.
Directive is implemented in the router repo.

Router part: https://github.com/wundergraph/cosmo/pull/2160